backend/mcp_servers/terminal/requirements.txt
# MCP Server
mcp>=0.9.0

# Ollama
ollama-python>=0.1.0

# Utilities
python-dotenv>=1.0.0

